OneDrive/Attachments/Desktop/Alert_Incident_Intelligence-master/chat-kenex/requirements.txt
langchain
langchain-core
langchain-community
langchain-pinecone
langchain-huggingface
huggingface-hub
sentence-transformers
pinecone
python-dotenv
pypdf
flask
flask-cors